+++ src/Text/LLVM.hs
@@ -281,6 +281,7 @@ define attrs rty fun sig k = do
     , defBody       = body
     , defMetadata   = Map.empty
     , defComdat     = Nothing
+    , defPersonality = Nothing
     }
 
 -- | A combination of define and @freshSymbol@.
@@ -310,6 +311,7 @@ define' attrs rty sym sig va k = do
     , defBody       = snd (runBB (k (map (fmap toValue) args)))
     , defMetadata   = Map.empty
     , defComdat     = Nothing
+    , defPersonality = Nothing
     }
 
 -- Basic Block Monad -----------------------------------------------------------
@@ -730,18 +732,18 @@ getelementptr ty ptr ixs = observe ty (GEP [] ty (toValue `fmap` ptr) ixs)
 -- | Emit a call instruction, and generate a new variable for its result.
 call :: IsValue a => Typed a -> [Typed Value] -> BB (Typed Value)
 call sym vs = case typedType sym of
-  PtrTo ty@(FunTy rty _ _) -> observe rty (Call False ty (toValue sym) vs)
+  PtrTo ty@(FunTy rty _ _) -> observe rty (Call False ty (toValue sym) vs [])
   _                        -> error "invalid function type given to call"
 
 -- | Emit a call instruction, but don't generate a new variable for its result.
 call_ :: IsValue a => Typed a -> [Typed Value] -> BB ()
-call_ sym vs = effect (Call False (typedType sym) (toValue sym) vs)
+call_ sym vs = effect (Call False (typedType sym) (toValue sym) vs [])
 
 -- | Emit an invoke instruction, and generate a new variable for its result.
 invoke :: IsValue a =>
           Type -> a -> [Typed Value] -> Ident -> Ident -> BB (Typed Value)
 invoke rty sym vs to uw = observe rty
-                        $ Invoke rty (toValue sym) vs (Named to) (Named uw)
+                        $ Invoke rty (toValue sym) vs (Named to) (Named uw) []
 
 -- | Emit a call instruction, but don't generate a new variable for its result.
 switch :: IsValue a => Typed a -> Ident -> [(Integer, Ident)] -> BB ()