+++ src/Text/LLVM/DebugUtils.hs
@@ -416,7 +416,7 @@ localVariableNameDeclarations mdMap def =
 
     aux :: [Stmt] -> Map Ident Ident -> Map Ident Ident
     aux ( Effect (Store src dst _ _) _ _
-        : Effect (Call _ _ (ValSymbol (Symbol what)) [var,md,_]) _ _
+        : Effect (Call _ _ (ValSymbol (Symbol what)) [var,md,_] _) _ _
         : _) sofar
       | what == "llvm.dbg.declare"  -- pre-LLVM19: intrinsic declaration match
       , Just dstIdent <- extractIdent dst
@@ -426,7 +426,7 @@ localVariableNameDeclarations mdMap def =
       , Just name <- extractLvName md
       = Map.insert name srcIdent sofar
 
-    aux ( Effect (Call _ _ (ValSymbol (Symbol what)) [var,_,md,_]) _ _
+    aux ( Effect (Call _ _ (ValSymbol (Symbol what)) [var,_,md,_] _) _ _
         : _) sofar
       | what == "llvm.dbg.value"  -- pre-LLVM19: intrinsic declaration match
       , Just key  <- extractIdent var