+++ src/Text/LLVM/Parser.hs
@@ -48,6 +48,7 @@ pPrimType = choice
   , try (string "label")    >> return Label
   , try (string "void")     >> return Void
   , try (string "x86mmx")   >> return X86mmx
+  , try (string "token")    >> return Token
   , try (string "metadata") >> return Metadata
   ]